Re-build models output Add MySQL to models output test

+++ tests/sample_models/orders/orders_ddl_mssql_version1.sql
@@ -59,7 +59,7 @@ CREATE TABLE item (
 	product_version VARCHAR(1000) NULL, 
 	note VARCHAR(1000) NULL, 
 	quantity INTEGER NULL, 
-	price FLOAT NULL, 
+	price DOUBLE PRECISION NULL, 
 	CONSTRAINT cx_pk_item PRIMARY KEY CLUSTERED (pk_item), 
 	FOREIGN KEY(fk_parent_shiporder) REFERENCES shiporder (pk_shiporder)
 )
